% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.r
\name{arithmetic_geometric_mean}
\alias{arithmetic_geometric_mean}
\alias{geometric_harmonic_mean}
\title{Iterated mean functions}
\usage{
arithmetic_geometric_mean(x, y)

geometric_harmonic_mean(x, y)
}
\arguments{
\item{x, y}{The numbers to calculate the mean of.}
}
\value{
The arithmetic-geometric mean.

The geometric-harmonic mean.
}
\description{
Calculate the iterated means of two numbers, using functions generated by the function factory \code{make_iterated_mean}.
}
\examples{
arithmetic_geometric_mean(1,2)
geometric_harmonic_mean(1,2)
}
\references{
Weisstein, Eric W. (accessed 2023) Arithmetic-Geometric Mean. From MathWorld--A Wolfram Web Resource. https://mathworld.wolfram.com/Arithmetic-GeometricMean.html

Weisstein, Eric W. (accessed 2023) Arithmetic-Harmonic Mean. From MathWorld--A Wolfram Web Resource. https://mathworld.wolfram.com/Arithmetic-HarmonicMean.html
}
\seealso{
Other iterated means: 
\code{\link{arithmetic_mean2}()},
\code{\link{make_iterated_mean}()}

Other bivariate means: 
\code{\link{Heronian_mean}()},
\code{\link{arithmetic_mean2}()},
\code{\link{identric_mean}()},
\code{\link{logarithmic_mean}()},
\code{\link{make_Heinz_mean}()},
\code{\link{make_Seiffert_like_mean}()},
\code{\link{make_Stolarsky_mean}()},
\code{\link{make_iterated_mean}()}
}
\concept{bivariate means}
\concept{iterated means}
