% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov.r
\name{CAGR}
\alias{CAGR}
\alias{weighted_CAGR}
\title{CAGR}
\usage{
CAGR(x, na.rm = FALSE)

weighted_CAGR(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An object consisting of a number of percentage changes over a number of years.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
The CAGR.
}
\description{
Obtain the compound annual growth rate (CAGR) of an object.
}
\examples{
CAGR(c(-10,60))

weighted_CAGR(c(-10,60),c(1,2))
}
\seealso{
Other Kolmogorov means: 
\code{\link{cubic_mean}()},
\code{\link{geometric_mean}()},
\code{\link{harmonic_mean}()},
\code{\link{log_semiring_mean}()},
\code{\link{make_Holder_mean}()},
\code{\link{make_Kolmogorov_mean}()},
\code{\link{quadratic_mean}()},
\code{\link{sqrt_mean}()},
\code{\link{weighted_min}()}
}
\concept{Kolmogorov means}
