% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modular.r
\name{circular_mean}
\alias{circular_mean}
\alias{modular_mean}
\alias{degrees_mean}
\alias{gradians_mean}
\alias{weighted_circular_mean}
\alias{weighted_modular_mean}
\alias{weighted_degrees_mean}
\alias{weighted_gradians_mean}
\title{Circular mean}
\usage{
circular_mean(x, na.rm = FALSE)

modular_mean(x, m = 1, na.rm = FALSE)

degrees_mean(x, na.rm = FALSE)

gradians_mean(x, na.rm = FALSE)

weighted_circular_mean(x, w, na.rm = FALSE)

weighted_modular_mean(x, w, m, na.rm = FALSE)

weighted_degrees_mean(x, w, na.rm = FALSE)

weighted_gradians_mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object that can support trigonometric functions, representing a series of angles measured in radians or number over a given modulus.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{m}{The modulus.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
The circular mean of these angles, between \code{-pi} and \code{+pi}.

The modular mean of these numbers, between 0 and \code{m}.
}
\description{
Obtain the circular mean of an object.
}
\details{
A danger of circular mean functions is that where data points are perfectly balanced around the unit circle, the circular mean becomes singular, in which case zero or effectively random results can be returned dependent upon rounding.
}
\examples{
circular_mean(1:5)

modular_mean(c(2,8),10)

weighted_circular_mean(1:5,c(0,1,1,1,0))

weighted_modular_mean(c(2,8,9),c(1,1,0),10)
}
