% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lehmer.r
\name{contraharmonic_mean}
\alias{contraharmonic_mean}
\alias{weighted_contraharmonic_mean}
\title{Contraharmonic mean}
\usage{
contraharmonic_mean(x, na.rm = FALSE)

weighted_contraharmonic_mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
The contraharmonic mean.
}
\description{
Obtain the contraharmonic mean of an an object, also known as the self-weighting mean. I.e. the sum of the squared data divided by the sum of the data. This function is derived from the function factory \code{make_Lehmer_mean} with \code{p=2}.
}
\examples{
contraharmonic_mean(1:3)

weighted_contraharmonic_mean(1:4,c(1,1,1,0))
}
\references{
Zelen, M. (1972) Length-biased sampling and biomedical problems. In Biometric Society Meeting, Dallas, Texas.
}
\seealso{
Other Lehmer means: 
\code{\link{make_Lehmer_mean}()}
}
\concept{Lehmer means}
