% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Frechet.r
\name{make_Frechet_mean}
\alias{make_Frechet_mean}
\alias{make_weighted_Frechet_mean}
\alias{Frechet_mean}
\alias{weighted_Frechet_mean}
\title{Fréchet mean functions}
\usage{
make_Frechet_mean(DIST, ...)

make_weighted_Frechet_mean(DIST, ...)

Frechet_mean(x, DIST, na.rm = FALSE, ...)

weighted_Frechet_mean(x, w, DIST, na.rm = FALSE, ...)
}
\arguments{
\item{DIST}{a function to scale the measure \eqn{x - \bar x}, which is required to be vectorised.}

\item{...}{Other parameters to be passed to \code{nlm}.}

\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
A function with parameters \verb{(x, na.rm=FALSE)} to calculate a Fréchet mean of an object. These functions will return \code{NA} if the data contain \code{NA} or \code{NaN} and \code{na.rm=FALSE}, otherwise \code{NA} and \code{NaN} values are removed prior to further computation.

A function with parameters \verb{(x, w, na.rm=FALSE)} to calculate a weighted Frechet mean of an object.

A Frechet mean.
}
\description{
Also known as a Karcher mean. Calculate a mean by minimising the sum of a distance metric from the mean to the data. For different choices of distance metric this generalizes means such as the arithmetic mean and the median.
}
\details{
For repeat use use associated factory function and apply it, rather than the single-function-call equivalent.
}
\examples{
x <- c(1,3,6,10,15) 
my_average_mean <- make_Frechet_mean(\(z) z*z)
my_average_mean(x)

my_average_median <- make_Frechet_mean(abs)
my_average_median(x)

Frechet_mean(c(1,3,6,10,15),abs)

weighted_Frechet_mean(c(1,3,6,10,15),c(1,1,1,0,0),abs)
}
\concept{Frechet means}
