% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov.r
\name{geometric_mean}
\alias{geometric_mean}
\alias{weighted_geometric_mean}
\alias{geometric_sd}
\alias{weighted_geometric_sd}
\title{Geometric mean}
\usage{
geometric_mean(x, na.rm = FALSE)

weighted_geometric_mean(x, w, na.rm = FALSE)

geometric_sd(x, na.rm = FALSE)

weighted_geometric_sd(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
The geometric mean.

The geometric standard deviation.
}
\description{
Obtain the geometric mean of an an object. This function is derived from the function factory \code{make_Kolmogorov_mean} using the transformation functions \code{FUN=log} and \code{INV=exp}.
}
\examples{
geometric_mean(c(1,10,100))

weighted_geometric_mean(c(1,10,100),c(0,1,1))

geometric_sd(c(1,10,100))
}
\seealso{
Other Kolmogorov means: 
\code{\link{CAGR}()},
\code{\link{cubic_mean}()},
\code{\link{harmonic_mean}()},
\code{\link{log_semiring_mean}()},
\code{\link{make_Holder_mean}()},
\code{\link{make_Kolmogorov_mean}()},
\code{\link{quadratic_mean}()},
\code{\link{sqrt_mean}()},
\code{\link{weighted_min}()}
}
\concept{Kolmogorov means}
