% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov.r
\name{harmonic_mean}
\alias{harmonic_mean}
\alias{weighted_harmonic_mean}
\title{Harmonic mean}
\usage{
harmonic_mean(x, na.rm = FALSE)

weighted_harmonic_mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
The harmonic mean.
}
\description{
Obtain the harmonic mean of an an object, also known as the subcontrary mean. This function is derived from the function factory \code{make_Kolmogorov_mean} using the transformation functions \verb{FUN=\\(z) 1/z} and \verb{INV=\\(z) 1/z}.
}
\examples{
harmonic_mean(c(1,10,100))

weighted_harmonic_mean(c(1,10,100),c(0,1,1))
}
\seealso{
Other Kolmogorov means: 
\code{\link{CAGR}()},
\code{\link{cubic_mean}()},
\code{\link{geometric_mean}()},
\code{\link{log_semiring_mean}()},
\code{\link{make_Holder_mean}()},
\code{\link{make_Kolmogorov_mean}()},
\code{\link{quadratic_mean}()},
\code{\link{sqrt_mean}()},
\code{\link{weighted_min}()}
}
\concept{Kolmogorov means}
