% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.r
\name{Heronian_mean}
\alias{Heronian_mean}
\title{Heronian mean}
\usage{
Heronian_mean(x, y)
}
\arguments{
\item{x, y}{The numbers to calculate the mean of.}
}
\value{
The Heronian mean.
}
\description{
Calculate the Heronian mean, which is for two numbers \code{x} and \code{y} equal to \eqn{\frac{1}{3}(x+y+\sqrt{xy})}.
}
\examples{
Heronian_mean(1,2)
}
\references{
Bullen, P.S. (2003). 2.1.4 Heronian, centroidal and neo-Pythagorean means, Handbook of Means and Their Inequalities. Mathematics and Its Applications, Berlin, New York: Springer Science+Business Media, pp. 399–401, doi:10.1007/978-94-017-0399-4, ISBN 978-1-4020-1522-9.
}
\seealso{
Other bivariate means: 
\code{\link{arithmetic_geometric_mean}()},
\code{\link{arithmetic_mean2}()},
\code{\link{identric_mean}()},
\code{\link{logarithmic_mean}()},
\code{\link{make_Heinz_mean}()},
\code{\link{make_Seiffert_like_mean}()},
\code{\link{make_Stolarsky_mean}()},
\code{\link{make_iterated_mean}()}
}
\concept{bivariate means}
