% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.r
\name{identric_mean}
\alias{identric_mean}
\title{Identric mean}
\usage{
identric_mean(x, y)
}
\arguments{
\item{x, y}{The numbers to calculate the mean of.}
}
\value{
The identric mean.
}
\description{
Calculate the identric mean of two numbers.
}
\examples{
identric_mean(1,2)
}
\references{
Alzer, H. (1993). Some Gamma Function Inequalities." Mathematics of Computation, 60, 337-346.

Bullen, P. S., Mitrinović, D. S., and Vasić, P. M. (1988). Means and Their Inequalities. Dordrecht, Netherlands: Reidel.
}
\seealso{
Other bivariate means: 
\code{\link{Heronian_mean}()},
\code{\link{arithmetic_geometric_mean}()},
\code{\link{arithmetic_mean2}()},
\code{\link{logarithmic_mean}()},
\code{\link{make_Heinz_mean}()},
\code{\link{make_Seiffert_like_mean}()},
\code{\link{make_Stolarsky_mean}()},
\code{\link{make_iterated_mean}()}
}
\concept{bivariate means}
