% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncated.r
\name{interquartile_mean}
\alias{interquartile_mean}
\title{Interquartile mean}
\usage{
interquartile_mean(x, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The interquartile mean.
}
\description{
Obtain the interquartile mean of an an object, which excludes the first and fourth quartiles of data. This function is derived from the function factory \code{make_truncated_mean}. Interpolation is specified where the length of \code{x} is not a multiple of 4.
}
\examples{
interquartile_mean((1:10)^2)
}
\seealso{
Other truncated means: 
\code{\link{Winsorized_mean}()},
\code{\link{make_trimmed_range}()},
\code{\link{make_truncated_mean}()},
\code{\link{modified_mean}()},
\code{\link{truncated_mean}()}
}
\concept{truncated means}
