% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov.r
\name{log_semiring_mean}
\alias{log_semiring_mean}
\alias{weighted_log_semiring_mean}
\alias{LogSumExp}
\alias{weighted_LogSumExp}
\title{Log semiring mean}
\usage{
log_semiring_mean(x, na.rm = FALSE)

weighted_log_semiring_mean(x, w, na.rm = FALSE)

LogSumExp(x, na.rm = FALSE)

weighted_LogSumExp(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
The log semiring mean.

The LogSumExp value of the object.

The weighted LogSumExp value of the object.
}
\description{
Obtain the log semiring mean of an an object. This function is derived from the function factory \code{make_Kolmogorov_mean} using the transformation functions \code{FUN=exp} and \code{INV=log}. This is the opposite mean to the geometric mean. The \code{LogSumExp} function is similar but uses \code{FUNC=sum} rather than \code{FUNC=mean}.
}
\examples{
log_semiring_mean(1:3)

weighted_log_semiring_mean(1:4,c(1,1,1,0))

LogSumExp(1:4)

weighted_LogSumExp(1:4,c(1,1,1,0))
}
\seealso{
Other Kolmogorov means: 
\code{\link{CAGR}()},
\code{\link{cubic_mean}()},
\code{\link{geometric_mean}()},
\code{\link{harmonic_mean}()},
\code{\link{make_Holder_mean}()},
\code{\link{make_Kolmogorov_mean}()},
\code{\link{quadratic_mean}()},
\code{\link{sqrt_mean}()},
\code{\link{weighted_min}()}
}
\concept{Kolmogorov means}
