% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.r
\name{logarithmic_mean}
\alias{logarithmic_mean}
\title{Logarithmic mean}
\usage{
logarithmic_mean(x, y)
}
\arguments{
\item{x, y}{The numbers to calculate the mean of.}
}
\value{
The logarithmic mean.
}
\description{
Calculate the logarithmic mean of two numbers, not to be confused with the geometric mean calculated by \code{geometric_mean2}.
}
\examples{
logarithmic_mean(1,2)
}
\references{
Tung-Po Lin (1974). The Power Mean and the Logarithmic Mean. The American Mathematical Monthly, 81: 879–883. doi:10.1080/00029890.1974.11993684.
}
\seealso{
Other bivariate means: 
\code{\link{Heronian_mean}()},
\code{\link{arithmetic_geometric_mean}()},
\code{\link{arithmetic_mean2}()},
\code{\link{identric_mean}()},
\code{\link{make_Heinz_mean}()},
\code{\link{make_Seiffert_like_mean}()},
\code{\link{make_Stolarsky_mean}()},
\code{\link{make_iterated_mean}()}
}
\concept{bivariate means}
