% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lestimators.r
\name{make_average_absolute_deviation}
\alias{make_average_absolute_deviation}
\alias{mean_absolute_deviation}
\alias{median_absolute_deviation}
\title{Average absolute deviation}
\usage{
make_average_absolute_deviation(CENT, FUNC, const = 1)

mean_absolute_deviation(x, na.rm = FALSE)

median_absolute_deviation(x, na.rm = FALSE)
}
\arguments{
\item{CENT}{A function to determine a centre point of the data.}

\item{FUNC}{A function to apply to the deviations from the centre point of the data.}

\item{const}{A normalising constant to provide an estimate of the sample standard deviation.}

\item{x}{An \code{R} object that can support the \code{quantile} function.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
A function to calculate the average absolute deivation on an object

The mean absolute deviation (mean of absolute deviations from mean, scaled by \code{sqrt(pi/2)})

The median absolute deviation (mean of absolute deviations from median, scaled by \code{1.4826})
}
\description{
Calculates an average absolute deviation function using a function factory.
}
\examples{
MAD <- make_average_absolute_deviation(mean,mean,sqrt(pi/2))
MAD(1:100)

mean_absolute_deviation(1:100)

median_absolute_deviation(1:100)
}
