% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscAverages.r
\name{make_Bayesian_average}
\alias{make_Bayesian_average}
\alias{Bayesian_average}
\title{Create Bayesian average function}
\usage{
make_Bayesian_average(C, m)

Bayesian_average(x, C, m, na.rm = FALSE)
}
\arguments{
\item{C}{The prior sample size, which does not need to be an integer.}

\item{m}{The prior mean.}

\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
A function to calculate a Bayesian average of an object.

A Bayesian average.
}
\description{
Create a function to calculate a Bayesian average using a function factory. This is a mean assuming a prior mean \code{m} with representative weight \code{C}, increasing the numerator by \code{Cm} and the denominator by \code{C}.
}
\details{
For repeat use use the associated factory function and apply it.
}
\examples{
make_Bayesian_average(8,10)

Bayesian_average(c(5,6,7),8,10)
Bayesian_average(c(NA),1,1,na.rm=TRUE)
}
