% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gini.r
\name{make_Gini_mean}
\alias{make_Gini_mean}
\alias{make_weighted_Gini_mean}
\alias{Gini_mean}
\alias{weighted_Gini_mean}
\title{Gini mean functions}
\usage{
make_Gini_mean(r, s)

make_weighted_Gini_mean(r, s)

Gini_mean(x, r, s, na.rm = FALSE)

weighted_Gini_mean(x, w, r, s, na.rm = FALSE)
}
\arguments{
\item{r, s}{The parameters of the mean.}

\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
A function with parameters \verb{(x, na.rm=FALSE)} to calculate a Gini mean of an object, except for the following cases. For \code{s==0} the \code{make_Holder_mean(r)} function is returned. For \code{r==1} the \code{make_Lehmer_mean(s+1)} function is returned. For \code{r==-1} the \code{make_Lehmer_mean(s)} function is returned. For \code{r==0} a special-case function is returned.

A function with parameters \verb{(x, w, na.rm=FALSE)} to calculate a weighted Gini mean of an object, with associated weights, except for the following cases. For \code{s==0} the \code{make_weighted_Holder_mean(r)} function is returned. For \code{r==1} the \code{make_weighted_Lehmer_mean(s+1)} function is returned. For \code{r==-1} the \code{make_weighted_Lehmer_mean(s)} function is returned. For \code{r==0} a special-case function is returned.

A Gini mean
}
\description{
Create a Gini mean function using a function factory. Gini means are a two-parameter family of means with parameters \code{r} and \code{s} which generalize the Lehmer and Holder mean families, interpolating between various means.
}
\details{
For very high or low values of \code{p} overflow or underflow during internal calculations may generate unexpected results. For repeat use use associated factory function and apply it, rather than the single-function-call equivalent. Removal of \code{NA} and \code{NaN} values is performed by \code{mean}, \code{min}, \code{max} or \code{sum}, therefore will remove any such values introduced by manipulating the data.
}
\examples{
contramean <- make_Gini_mean(1,1)
contramean(c(1,2,3))

contrameanw <- make_weighted_Gini_mean(1,1)
contrameanw(c(1,2,3,4),c(1,1,1,0))

Gini_mean(c(1,2,3,4),1,1)
weighted_Gini_mean(c(1,2,3,4,5),c(1,1,1,1,0),1,1)
}
\references{
Farnsworth, D and Orr, R (1986). Gini means. The American Mathematical Monthly 93: 603-607.
}
\concept{Gini means}
