% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.r
\name{make_Heinz_mean}
\alias{make_Heinz_mean}
\alias{Heinz_mean}
\title{Heinz mean}
\usage{
make_Heinz_mean(p)

Heinz_mean(x, y, p)
}
\arguments{
\item{p}{The power for the mean.}

\item{x, y}{The numbers to calculate the mean of.}
}
\value{
A function with parameters \verb{(x,y)} to calculate a Heinz mean, except for the following cases. For \code{p==0} the \code{arithmetic_mean2} function is returned. For \code{p==0.5} the \code{geometric_mean2} function is returned.

A Heinz mean.
}
\description{
Create a Heinz mean function to apply to two numbers, using a function factory. For a power \code{p}, the Heinz mean interpolates between the arithmetic (\code{p==0}) and geometric means (\code{p==0.5}) for two numbers.
}
\details{
For repeat use use associated factory function and apply it, rather than the single-function-call equivalent.
}
\examples{
hm <- make_Heinz_mean(0.25)
hm(1,2)

Heinz_mean(1,2,0.25)
}
\references{
Bhatia, R. (2006). Interpolating the arithmetic-geometric mean inequality and its operator version. Linear Algebra and Its Applications, 413: 355–363. doi:10.1016/j.laa.2005.03.005.
}
\seealso{
Other bivariate means: 
\code{\link{Heronian_mean}()},
\code{\link{arithmetic_geometric_mean}()},
\code{\link{arithmetic_mean2}()},
\code{\link{identric_mean}()},
\code{\link{logarithmic_mean}()},
\code{\link{make_Seiffert_like_mean}()},
\code{\link{make_Stolarsky_mean}()},
\code{\link{make_iterated_mean}()}
}
\concept{bivariate means}
