% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov.r
\name{make_Holder_mean}
\alias{make_Holder_mean}
\alias{make_weighted_Holder_mean}
\alias{Holder_mean}
\alias{weighted_Holder_mean}
\title{Holder mean functions}
\usage{
make_Holder_mean(p)

make_weighted_Holder_mean(p)

Holder_mean(x, p, na.rm = FALSE)

weighted_Holder_mean(x, w, p, na.rm = FALSE)
}
\arguments{
\item{p}{The power for the transformation.}

\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
A function with parameters \verb{(x, na.rm=FALSE)} to calculate a Holder mean of an object, derived from \code{make_Kolmogorov_mean} except for the following cases. For \code{p==-Inf} the \code{min} function is returned. For \code{p==-1} the \code{harmonic_mean} function is returned. For \code{p==0} the \code{geometric_mean} function is returned. For \code{p==0.5} the \code{sqrt_mean} function is returned. For \code{p==1} the \code{mean} function is returned. For \code{p==2} the \code{quadratic_mean} function is returned. For \code{p==3} the \code{cubic_mean} function is returned. For \code{p==Inf} the \code{max} function is returned.

A function with parameters \verb{(x, w, na.rm=FALSE)} to calculate a weighted Holder mean of an object, derived from \code{make_weighted_Kolmogorov_mean} except for the following cases. For \code{p==-Inf} the \code{weighted_min} function is returned. For \code{p==-1} the \code{weighted_harmonic_mean} function is returned. For \code{p==0} the \code{weighted_geometric_mean} function is returned. For \code{p==0.5} the \code{weighted_sqrt_mean} function is returned. For \code{p==1} the \code{weighted_mean} function is returned. For \code{p==2} the \code{weighted_quadratic_mean} function is returned. For \code{p==3} the \code{weighted_cubic_mean} function is returned. For \code{p==Inf} the \code{weighted_max} function is returned.

The Holder mean, calculated by a function returned by \code{make_Holder_mean}.

The weighted Holder mean, calculated by a function returned by \code{make_weighted_Holder_mean}.
}
\description{
Create a Holder mean function using a function factory. Also known as a power mean. This function is derived from the function factory \code{make_Kolmogorov_mean} using the transformation functions \verb{FUN=\\(z) z^p} and \verb{INV=\\(z) z^(1/p)} except for certain values of \code{p} where predefined functions are returned instead.
}
\details{
For very high or low values of \code{p} overflow or underflow during internal calculations may generate unexpected results. For repeat use use associated factory function and apply it, rather than the single-function-call equivalent.
}
\examples{
geomean <- make_Holder_mean(0)
geomean(c(1,10,100))

geomeanw <- make_weighted_Holder_mean(0)
geomeanw(c(1,10,100,1000),c(1,1,1,0))

Holder_mean(c(1,10,100),0)

weighted_Holder_mean(c(1,10,100,1000),c(1,1,1,0),0)
}
\seealso{
Other Kolmogorov means: 
\code{\link{CAGR}()},
\code{\link{cubic_mean}()},
\code{\link{geometric_mean}()},
\code{\link{harmonic_mean}()},
\code{\link{log_semiring_mean}()},
\code{\link{make_Kolmogorov_mean}()},
\code{\link{quadratic_mean}()},
\code{\link{sqrt_mean}()},
\code{\link{weighted_min}()}
}
\concept{Kolmogorov means}
