% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.r
\name{make_iterated_mean}
\alias{make_iterated_mean}
\alias{iterated_mean}
\title{Create iterated mean function}
\usage{
make_iterated_mean(FUN1, FUN2, tol = .Machine$double.eps^0.5)

iterated_mean(x, y, FUN1, FUN2, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{FUN1, FUN2}{The functions to apply.}

\item{tol}{The tolerance required before convergence is considered achieved.}

\item{x, y}{The numbers to calculate the mean of.}
}
\value{
A function with parameters \verb{(x, y)} to calculate an iterated mean.

The iterated mean.
}
\description{
Create an iterated mean function to apply to two numbers, using a function factory. For an iterated mean, two functions \code{FUN1} and \code{FUN2} are applied to the pair of numbers, producing a new pair of numbers, and the process repeated until convergence.
}
\details{
For repeat use use associated factory function and apply it, rather than the single-function-call equivalent.
}
\examples{
agmean <- make_iterated_mean(\(i,j) (i+j)/2, \(i,j) sqrt(i*j))
agmean(1,2)

iterated_mean(iterated_mean(1 ,2, \(i,j) (i+j)/2, \(i,j) sqrt(i*j)))
}
\seealso{
Other iterated means: 
\code{\link{arithmetic_geometric_mean}()},
\code{\link{arithmetic_mean2}()}

Other bivariate means: 
\code{\link{Heronian_mean}()},
\code{\link{arithmetic_geometric_mean}()},
\code{\link{arithmetic_mean2}()},
\code{\link{identric_mean}()},
\code{\link{logarithmic_mean}()},
\code{\link{make_Heinz_mean}()},
\code{\link{make_Seiffert_like_mean}()},
\code{\link{make_Stolarsky_mean}()}
}
\concept{bivariate means}
\concept{iterated means}
