% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov.r
\name{make_Kolmogorov_mean}
\alias{make_Kolmogorov_mean}
\alias{make_weighted_Kolmogorov_mean}
\alias{Kolmogorov_mean}
\alias{weighted_Kolmogorov_mean}
\title{Kolmogorov mean functions}
\usage{
make_Kolmogorov_mean(TRA, INV, FUNC = mean)

make_weighted_Kolmogorov_mean(TRA, INV, FUNC = weighted_mean)

Kolmogorov_mean(x, TRA, INV, na.rm = FALSE)

weighted_Kolmogorov_mean(x, w, TRA, INV, na.rm = FALSE)
}
\arguments{
\item{TRA}{The transformation function.}

\item{INV}{The inverse function corresponding to TRA.}

\item{FUNC}{The actual function to apply to the transformed data, able to support \code{na.rm}.}

\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
A function with parameters \verb{(x, na.rm=FALSE)} to calculate a Kolmogorov mean of an object.

A function with parameters \verb{(x, w, na.rm=FALSE)} to calculate a weighted Kolmogorov mean of an object.

A Kolmogorov mean of an object.
}
\description{
Create a quasi-arithmetic mean function using a function factory. Also known as a Kolmogorov mean or generalised \emph{f}-mean. A summary statistic is calculated using the generic \code{mean} or other supplied function \code{FUNC} on transformed data using the function \code{TRA} and back transformed using the function \code{INV}, which are required to be vectorised.
}
\details{
For repeat use use associated factory function and apply it, rather than the single-function-call equivalent. Removal of \code{NA} and \code{NaN} values is performed by \code{mean}, therefore will remove any such values introduced by \code{TRA}.
}
\examples{
geomean <- make_Kolmogorov_mean(log,exp)
geomean(c(1,10,100))
geomean(c(1,10,100,NA), na.rm=TRUE)

geomeanw <- make_weighted_Kolmogorov_mean(log,exp)
geomeanw(c(1,10,100),c(0,1,1))

Kolmogorov_mean(c(1,10,100),log,exp)

weighted_Kolmogorov mean(c(1,10,100),c(0,1,1),log,exp)
}
\seealso{
Other Kolmogorov means: 
\code{\link{CAGR}()},
\code{\link{cubic_mean}()},
\code{\link{geometric_mean}()},
\code{\link{harmonic_mean}()},
\code{\link{log_semiring_mean}()},
\code{\link{make_Holder_mean}()},
\code{\link{quadratic_mean}()},
\code{\link{sqrt_mean}()},
\code{\link{weighted_min}()}
}
\concept{Kolmogorov means}
