% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lehmer.r
\name{make_Lehmer_mean}
\alias{make_Lehmer_mean}
\alias{make_weighted_Lehmer_mean}
\alias{Lehmer_mean}
\alias{weighted_Lehmer_mean}
\title{Lehmer mean functions}
\usage{
make_Lehmer_mean(p)

make_weighted_Lehmer_mean(p)

Lehmer_mean(x, p, na.rm = FALSE)

weighted_Lehmer_mean(x, w, p, na.rm = FALSE)
}
\arguments{
\item{p}{The power used in the numerator of the expression.}

\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
A function with parameters \verb{(x, na.rm=FALSE)} to calculate a Lehmer mean of an object, except for the following cases. For \code{p==-Inf} the \code{min} function is returned. For \code{p==0} the \code{harmonic_mean} function is returned. For \code{p==1} the \code{mean} function is returned. For \code{p==2} a function optimised to avoid power functions is returned. For \code{p==Inf} the \code{max} function is returned.

A function with parameters \verb{(x, w, na.rm=FALSE)} to calculate a weighted Lehmer mean of an object, with associated weights, except for the following cases. For \code{p==-Inf} the \code{weighted_min} function is returned. For \code{p==0} the \code{weighted_harmonic_mean} function is returned. For \code{p==1} the \code{weighted_mean} function is returned. For \code{p==2} a function optimised to avoid power functions is returned. For \code{p==Inf} the \code{weighted_max} function is returned.

The Lehmer mean.
}
\description{
Create a Lehmer mean function using a function factory. For a power \code{p}, the Lehmer mean is the ratio of the sum of the data points raised to the power of \code{p} and the sum of the data points raised to the power of \code{p-1}. The mean interpolates between the arithmetic mean for \code{p=1} and the harmonic mean for \code{p=0} and includes other means at other values of \code{p}.
}
\details{
For very high or low values of \code{p} overflow or underflow during internal calculations may generate unexpected results. For repeat use use associated factory function and apply it, rather than the single-function-call equivalent. Removal of \code{NA} and \code{NaN} values is performed by \code{mean}, \code{min}, \code{max} or \code{sum}, therefore will remove any such values introduced by manipulating the data.
}
\examples{
contramean <- make_Lehmer_mean(2)
contramean(c(1,2,3))

contrameanw <- make_weighted_Lehmer_mean(2)
contrameanw(c(1,2,3,4),c(1,1,1,0))

Lehmer_mean(c(1,2,3),2)

weighted_Lehmer_mean(c(1,2,3,4),c(1,1,1,0),2)
}
\references{
Bullen, P. S. (1987). Handbook of means and their inequalities. Springer.
}
\seealso{
Other Lehmer means: 
\code{\link{contraharmonic_mean}()}
}
\concept{Lehmer means}
