% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lestimators.r
\name{make_midsummary}
\alias{make_midsummary}
\alias{midsummary}
\alias{midhinge}
\alias{midrange}
\title{Create midsummary function}
\usage{
make_midsummary(p)

midsummary(x, p, na.rm = FALSE)

midhinge(x, na.rm = FALSE)

midrange(x, na.rm = FALSE)
}
\arguments{
\item{p}{The fraction (0 to 0.5) of observations to be trimmed from each end of the data before the mean is computed.}

\item{x}{An \code{R} object that can support the \code{quantile} function.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
A function to calculate the midsummary on an object, produced by \code{n_point_L_estimator}.

The midsummary statistic.

The midhinge statistic.

The midrange statistic.
}
\description{
Create a midsummary function using a function factory.
}
\details{
For repeat use use associated factory function and apply it, rather than the single-function-call equivalent.
}
\examples{
my_midsummary <- make_midsummary(0.1)
my_midsummary((1:10)^2)

midsummary((1:10)^2,0.1)
midhinge((1:10)^2)
midrange((1:10)^2)
}
\seealso{
Other L estimators: 
\code{\link{make_trimmed_range}()},
\code{\link{n_point_L_estimator}()},
\code{\link{trimean}()}
}
\concept{L estimators}
