% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bivariate.r
\name{make_Seiffert_like_mean}
\alias{make_Seiffert_like_mean}
\alias{Seiffert_like_mean}
\alias{Neuman_Sandor}
\alias{first_Seiffert}
\alias{second_Seiffert}
\title{Seiffert-like mean}
\usage{
make_Seiffert_like_mean(FUN)

Seiffert_like_mean(x, y, FUN)

Neuman_Sandor(x, y)

first_Seiffert(x, y)

second_Seiffert(x, y)
}
\arguments{
\item{FUN}{The positive function for the mean.}

\item{x, y}{The numbers to calculate the mean of.}
}
\value{
A function with parameters \verb{(x,y)} to calculate a Seiffert-like mean.

A Seiffert-like mean.
}
\description{
Create a Seiffert-like mean function to apply to two numbers, using a function factory. For a function \code{FUN}, the Seiffert-like means include functions such as the arithmetic mean (\code{FUN=identity}), Neuman-Sandor mean (\code{FUN=asinh}), logarithmic mean (\code{FUN=atanh}) and Seiffert means (\code{FUN=asin} and \code{FUN=atan}).
}
\details{
For repeat use use associated factory function and apply it, rather than the single-function-call equivalent.
}
\examples{
ns <- make_Seiffert_like_mean(asinh)
ns(1,2)

Seiffert_like_mean(1,2,asinh)
Neuman_Sandor(1,2)

first_Seiffert(1,2)

second_Seiffert(1,2)
}
\references{
Neuman, E & Sándor, J (2003). On the Schwab–Borchardt mean, Mathetica Pannonica 14: 253–266.

Witkowski, A (2015). On Seiffert-like means. Journal of Mathematical Inequalities 9: 1071-1092.
}
\seealso{
Other bivariate means: 
\code{\link{Heronian_mean}()},
\code{\link{arithmetic_geometric_mean}()},
\code{\link{arithmetic_mean2}()},
\code{\link{identric_mean}()},
\code{\link{logarithmic_mean}()},
\code{\link{make_Heinz_mean}()},
\code{\link{make_Stolarsky_mean}()},
\code{\link{make_iterated_mean}()}
}
\concept{bivariate means}
