% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lestimators.r
\name{make_trimmed_range}
\alias{make_trimmed_range}
\alias{trimmed_range}
\alias{interdecile_range}
\alias{interquartile_range}
\title{Trimmed range}
\usage{
make_trimmed_range(p)

trimmed_range(x, p, na.rm = FALSE)

interdecile_range(x, na.rm = FALSE)

interquartile_range(x, na.rm = FALSE)
}
\arguments{
\item{p}{The fraction (0 to 0.5) of observations to be trimmed from each end of the data before the range is computed.}

\item{x}{An \code{R} object that can support the \code{quantile} function.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
A function to calculate the trimmed range on an object

The trimmed range.

The interdecile range.

The interquartile range.
}
\description{
Calculates a trimmed range function using a function factory.
}
\details{
For repeat use use associated factory function and apply it, rather than the single-function-call equivalent.
}
\examples{
iqr <- make_trimmed_range(0.25)
iqr(1:100)

trimmed_range(1:100,0.1)

interdecile_range(1:100)

interquartile_range(1:100)
}
\seealso{
Other L estimators: 
\code{\link{make_midsummary}()},
\code{\link{n_point_L_estimator}()},
\code{\link{trimean}()}

Other truncated means: 
\code{\link{Winsorized_mean}()},
\code{\link{interquartile_mean}()},
\code{\link{make_truncated_mean}()},
\code{\link{modified_mean}()},
\code{\link{truncated_mean}()}
}
\concept{L estimators}
\concept{truncated means}
