% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncated.r
\name{make_truncated_mean}
\alias{make_truncated_mean}
\title{Create truncated or Winsorized mean function}
\usage{
make_truncated_mean(
  FUNC,
  tail,
  type = "count",
  Winsorize = FALSE,
  interpolate = TRUE
)
}
\arguments{
\item{FUNC}{The function to apply to the truncated or Winsorized data, usually \code{mean}.}

\item{tail}{The size of tail at each end of the sorted data to be truncated or Winsorized.}

\item{type}{The type of tail. Can be \code{count}, \code{quartile} (0-4), \code{quantile} (0-1), \code{decile} (0-10), or \code{percentile} (0-100).}

\item{Winsorize}{The tails are Winsorized if true, otherwise excluded.}

\item{interpolate}{Tails are conservative/exclusive if false, otherwise interpolation is performed.}
}
\value{
A function with parameters \verb{(x, na.rm=FALSE)} to calculate a truncated or Winsorized mean on an object. These functions will return \code{NA} if the data contain \code{NA} or \code{NaN} and \code{na.rm=FALSE}, otherwise \code{NA} and \code{NaN} values are removed prior to further computation.
}
\description{
Create a mean function allowing for either truncation or Winsorizing of data, including interpolation, using a function factory. For a truncated mean, a size of tail is specified and data are excluded. For a Winsorized mean, a size of tail is specified and data in the tails are replaced by the nearest data point. Where interpolation is specified, means are interpolated between the nearest possible data point to the specified quantile, counting inclusively and exclusively.
}
\examples{
iqmean <- make_truncated_mean(mean, 1, type="quartile", interpolate=TRUE)
iqmean((1:10)^2)
}
\seealso{
Other truncated means: 
\code{\link{Winsorized_mean}()},
\code{\link{interquartile_mean}()},
\code{\link{make_trimmed_range}()},
\code{\link{modified_mean}()},
\code{\link{truncated_mean}()}
}
\concept{truncated means}
