% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PinkAverage-package.r
\docType{package}
\name{PinkAverage}
\alias{PinkAverage}
\title{PinkAverage}
\description{
Calculate various different types of averages and other related summary statistics. Individual functions for various types of means or averages are included, and function factories for creating families of means such as  Gini, Holder, and Lehmer means, both weighted and unweighted.
}
\section{Author}{

\itemize{
\item Darren Green \href{mailto:darren.green@stir.ac.uk}{darren.green@stir.ac.uk}
}
}

\section{Function factories}{

For most families of means, four functions are included. Function factories are preceded by \code{make}, e.g.

\code{f <- make_Susan_mean(params)}

\code{f(x)}

and

\code{f <- make_weighted_Susan_mean(params)}

\code{f(x,w)}.

For one-off use, equivalent functions to both make and then use the function factories are provided as

\code{Susan_mean(x, params)} and

\code{weighted_Susan_mean(x, w, params)}.

For repeat use use associated factory function in advance and apply it, rather than the single-function-call equivalent, as this is more efficient. Most factory-produced functions support the argument \code{na.rm} to remove \code{NA} and \code{NaN} values before computation proceeds.
}

\section{Families}{

Several families of means are implemented, including
\itemize{
\item The very general Kolmogorov mean family, including...
\item The power mean or Holder mean family.
\item The Lehmer mean family.
\item The Gini mean family, which genereralises the Holder and Lehmer families.
\item A family of truncated (including Winsorized) means.
\item A family of n-point L estimators including midsummary functions.
\item A family of Frechet (Karcher) means.
\item A family of circular and modular means.
\item A family of Bayesian averages.
}
}

\section{Bivariate means}{

Several families of means are implemented which operate on pairs of numbers rather than more general vectors of numbers, as follows:

\code{f <- make_Susan_mean(params)}

\code{f(x,y)}.
\itemize{
\item A factory-function family for iterated means.
\item A family of Seiffert-like means.
\item A family of Stolarsky means, including the identric.
\item A family of Heinz means.
}
}

\section{Miscellaneous means}{

Called as follows:

\code{Susan_mean(x,params)}
\itemize{
\item The mode.
\item The pseudomedian.
}
}

\section{Miscellaneous bivariate means}{

Called as follows:

\code{Susan_mean(x,y)}
\itemize{
\item The Heronian mean.
}
}

