% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiscAverages.r
\name{pseudomedian}
\alias{pseudomedian}
\title{Pseudomedian}
\usage{
pseudomedian(x, na.rm = FALSE, diag = FALSE, FUNC = median)
}
\arguments{
\item{x}{An \code{R} object that can support basic arithmetic.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{diag}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether pairs \verb{(i,i)} should be included.}

\item{FUNC}{The function to apply to the Walsh averages.}
}
\value{
The pseudomedian of these data.
}
\description{
Obtain the pseudomedian of an object. This is the median of the set midpoints of all pairwise combinations of data within the sample, these midpoints being the Walsh averages.
}
\details{
This implementation includes pairs consisting of the same data point twice.
}
\examples{
pseudo_median((1:10)^2)
}
