% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kolmogorov.r
\name{quadratic_mean}
\alias{quadratic_mean}
\alias{weighted_quadratic_mean}
\title{Quadratic mean}
\usage{
quadratic_mean(x, na.rm = FALSE)

weighted_quadratic_mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}
}
\value{
The quadratic mean.
}
\description{
Obtain the quadratic mean of an an object, also known as the root mean square or RMS. This function is derived from the function factory \code{make_Kolmogorov_mean} using the transformation functions \verb{FUN=\\(z) z*z} and \code{INV=sqrt}.
}
\examples{
quadratic_mean(1:3)

weighted_quadratic_mean(1:4,c(1,1,1,0))
}
\seealso{
Other Kolmogorov means: 
\code{\link{CAGR}()},
\code{\link{cubic_mean}()},
\code{\link{geometric_mean}()},
\code{\link{harmonic_mean}()},
\code{\link{log_semiring_mean}()},
\code{\link{make_Holder_mean}()},
\code{\link{make_Kolmogorov_mean}()},
\code{\link{sqrt_mean}()},
\code{\link{weighted_min}()}
}
\concept{Kolmogorov means}
