% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lestimators.r
\name{trimean}
\alias{trimean}
\alias{Swansons_rule}
\title{Trimean}
\usage{
trimean(x, na.rm = FALSE)

Swansons_rule(x, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object that can support the \code{quantile} function.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The trimean statistic.

The Swanson's rule statistic.
}
\description{
Obtain the trimean of an object.
}
\examples{
trimean((1:10)^2)
Swansons_rule((1:10)^2)
}
\seealso{
Other L estimators: 
\code{\link{make_midsummary}()},
\code{\link{make_trimmed_range}()},
\code{\link{n_point_L_estimator}()}
}
\concept{L estimators}
