% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncated.r
\name{truncated_mean}
\alias{truncated_mean}
\title{Truncated mean}
\usage{
truncated_mean(x, tail, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{tail}{The tails of the distribution to be removed, specified as a quantile (0-1).}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The truncated mean.
}
\description{
Obtain the truncated mean of an an object, which excludes the tails of the data. This function is derived from the function factory \code{make_truncated_mean}. Interpolation is specified where the length of \code{x} is not a multiple of 4.
}
\examples{
truncated_mean((1:10)^2,0.1)
}
\seealso{
Other truncated means: 
\code{\link{Winsorized_mean}()},
\code{\link{interquartile_mean}()},
\code{\link{make_trimmed_range}()},
\code{\link{make_truncated_mean}()},
\code{\link{modified_mean}()}
}
\concept{truncated means}
