% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.r
\name{weighted_mean}
\alias{weighted_mean}
\title{Weighted mean function}
\usage{
weighted_mean(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The weighted mean.
}
\description{
The core weighted.mean function treats zero-weighted infinities as zero. For internal consistency within the library, \code{weighted_mean} instead allows NaNs to propagate, treating the result of \code{0*Inf} to be singular.
}
\examples{
weighted_mean(c(-Inf,1,Inf),c(0,1,0))
weighted.mean(c(-Inf,1,Inf),c(0,1,0))
}
\seealso{
Other Weighted summary statistics: 
\code{\link{weighted_var}()}
}
\concept{Weighted summary statistics}
