% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.r
\name{weighted_min}
\alias{weighted_min}
\alias{weighted_max}
\title{Weighted minimum and maximum}
\usage{
weighted_min(x, w, na.rm = FALSE)

weighted_max(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The minimum value, or \code{NaN} where this has zero weighting.

The maximum value, or \code{NaN} where this has zero weighting.
}
\description{
Generates the minimum or maximum of an object, of those values where the weightings provided by a second object are non zero. These functions are returned as an asymptotic case for weighted mean factory functions where \code{min} and \code{max} would be returned by the non-weighted factory functions.
}
\details{
For a mean such as the Holder mean with parameter \eqn{p \to  -\infty}, the mean approaches the minimum (and vice versa for the maximum) value of the dataset for all non-zero weightings. For other means, the lowest value of the dataset does not contribute to the mean where its waiting zero. The current function treats this situation as a singularity and returns \code{NaN} where the minimum value has zero weights (and vice versa for the maximum).
}
\examples{
weighted_min(1:4,c(0,1,1,1))

weighted_max(1:4,c(1,1,1,0))
}
\seealso{
Other Kolmogorov means: 
\code{\link{CAGR}()},
\code{\link{cubic_mean}()},
\code{\link{geometric_mean}()},
\code{\link{harmonic_mean}()},
\code{\link{log_semiring_mean}()},
\code{\link{make_Holder_mean}()},
\code{\link{make_Kolmogorov_mean}()},
\code{\link{quadratic_mean}()},
\code{\link{sqrt_mean}()}
}
\concept{Kolmogorov means}
