% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.r
\name{weighted_var}
\alias{weighted_var}
\alias{weighted_sd}
\title{weighted variance}
\usage{
weighted_var(x, w, na.rm = FALSE, norm = TRUE)

weighted_sd(x, w, na.rm = FALSE, norm = TRUE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{w}{An \code{R} object of the same length as \code{x} representing a series of weights.}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}

\item{norm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether weights should be normalised to \code{n-1} before computation proceeds.}
}
\value{
The weighted variance.

The weighted standard deviation.
}
\description{
The weighted_var function treats zero-weighted allows NaNs to propagate, treating the result of \code{0*Inf} to be singular.
}
\details{
There are two forms of weighted variance: probability/reliability weights, and occurrences/repeat weights. By default, \code{weighted_variance} assumes the former and rescales the weights back to a total of \code{n-1}.
}
\examples{
weighted_var(c(1,2,3),c(0.5,0.5,0))
weighted_var(c(1,2,3),c(1,1,0))

weighted_sd(c(1,2,3),c(0.5,0.5,0))
}
\seealso{
Other Weighted summary statistics: 
\code{\link{weighted_mean}()}
}
\concept{Weighted summary statistics}
