% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Truncated.r
\name{Winsorized_mean}
\alias{Winsorized_mean}
\alias{Winsorised_mean}
\title{Winsorized mean}
\usage{
Winsorized_mean(x, tail, na.rm = FALSE)

Winsorised_mean(x, tail, na.rm = FALSE)
}
\arguments{
\item{x}{An \code{R} object.}

\item{tail}{The tails of the distribution to be Winsorized, specified as a quantile (0-1).}

\item{na.rm}{A logical evaluating to \code{TRUE} or \code{FALSE} indicating whether \code{NA} values should be stripped before the computation proceeds.}
}
\value{
The Winsorized mean.
}
\description{
Obtain the Winsorized mean of an an object, which replaces the tails of the data with the nearest value outside the tails. This function is derived from the function factory \code{make_truncated_mean}. Interpolation is specified where the length of \code{x} is not a multiple of 4.
}
\examples{
Winsorized_mean((1:10)^2,0.1)
Winsorised_mean((1:10)^2,0.1)
}
\seealso{
Other truncated means: 
\code{\link{interquartile_mean}()},
\code{\link{make_trimmed_range}()},
\code{\link{make_truncated_mean}()},
\code{\link{modified_mean}()},
\code{\link{truncated_mean}()}
}
\concept{truncated means}
